chore: quick fixes - Move App to top - update dev server
